Pre_labeling_tool/src/lambda/code/prelabeling_execute_preannotation_jobs_mapstate/requirements.txt
fuzzywuzzy
python-Levenshtein
amazon-textract-caller
numpy